django_backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
imagekitio==1.2.0
google-generativeai==0.8.3
django-cors-headers==4.3.1
Pillow==10.4.0
setuptools>=65.0.0